+++ components/src/components/BaseModal/BaseModal.styled.ts
@@ -1,5 +1,5 @@
 import styled from "@emotion/styled";
-import { ModalLayoutProps } from "@/types";
+import { ModalLayoutProps } from "@/types/modal";
 import {
   getModalPositionStyle,
   getModalSizeStyle,
