+++ components/src/lib/Modal/index.tsx
@@ -2,7 +2,7 @@ import { PropsWithChildren } from "react";
 import useOutsideClickRef from "@hooks/useOutsideClickRef";
 import useKeyDown from "@/hooks/useKeyDown";
 import BaseModal from "@/components/BaseModal";
-import { ModalLayoutProps, ModalDefaultProps } from "../../types";
+import { ModalLayoutProps, ModalDefaultProps } from "../../types/modal";
 
 interface ModalProps
   extends Omit<ModalDefaultProps, "size">,
