+++ hooks/src/lib/hooks/CardBrand/index.ts
@@ -1,7 +1,9 @@
 import { useMemo } from "react";
-import { getCardBrand, getCardBrandError } from "./utils";
+import { UseCardBrandReturn } from "../types";
+import { getCardBrand } from "@/lib/cardBrand/utils/getCardBrand";
+import { getCardBrandError } from "@/lib/cardBrand/utils/getCardBrandError";
 
-const useCardBrand = (cardNumber: string) => {
+const useCardBrand = (cardNumber: string): UseCardBrandReturn => {
   const cardBrand = useMemo(() => getCardBrand(cardNumber), [cardNumber]);
   const errorState = useMemo(() => getCardBrandError(cardBrand), [cardBrand]);
