--- hooks/src/lib/hooks/CardCVC/constants/index.ts
@@ -1,1 +1,0 @@
-export const INITIAL_CVC_STATE = { value: "" };
