+++ hooks/src/lib/hooks/CardCVC/utils/index.ts
@@ -1,5 +1,5 @@
 import { commonConditions, getErrorByRules } from "@utils/validation";
-import { ErrorState, ValidationRule } from "@/types";
+import { ErrorState, ValidationRule } from "@/types/validation";
 
 const cvcValidationRules: ValidationRule<string>[] = [
   {
