--- hooks/src/lib/hooks/CardExpiryDate/constants/index.ts
@@ -1,1 +1,0 @@
-export const INITIAL_EXPIRY_DATE_STATE = { value: "" };
