--- hooks/src/lib/hooks/CardNumber/constants/index.ts
@@ -1,3 +1,0 @@
-export const INITIAL_CARD_NUMBER_STATE = {
-  value: "",
-};
