--- hooks/src/lib/hooks/CardNumber/types/index.ts
@@ -1,4 +1,0 @@
-import { INITIAL_CARD_NUMBER_STATE } from "../constants";
-
-export type CardNumberState = typeof INITIAL_CARD_NUMBER_STATE;
-export type CardNumberKey = keyof CardNumberState;
