--- hooks/src/lib/hooks/CardPassword/constants/index.ts
@@ -1,1 +1,0 @@
-export const INITIAL_CARD_PASSWORD = { value: "" };
