+++ hooks/src/lib/index.ts
@@ -1,5 +1,6 @@
 export { default as useCardPassword } from "./hooks/CardPassword";
 export { default as useCardNumber } from "./hooks/CardNumber";
 export { default as useCardExpiryDate } from "./hooks/CardExpiryDate";
-export { default as useCardCvc } from "./hooks/CardCVC";
+export { default as useCardCVC } from "./hooks/CardCVC";
 export { default as useCardBrand } from "./hooks/CardBrand";
+export { default as useFormatCardNumber } from "./hooks/FormatCardNumber";
