+++ hooks/src/types/input.ts
@@ -1,0 +1,3 @@
+export interface BaseInputState {
+  value: string;
+}
